Enforce -Werror on our own sources (wrapper and example)
Add an ENABLE_WERROR option and apply a strict warning set plus -Werror to our own code only: the src/ wrapper (per-source, so the bundled Opus C keeps its relaxed flags) and the opus_to_wav example. The CI build job sets -DENABLE_WERROR=ON.

-Wno-error=maybe-uninitialized is gated to GCC: it covers a GCC false positive at -O2, but Clang rejects the unknown option under -Werror.

Fix the conversions the gate surfaced (explicit casts), including read_le32, where shifting a promoted uint8_t into the sign bit was signed-overflow UB.

+++ host_examples/opus_to_wav/opus_to_wav.cpp
@@ -93,7 +93,7 @@ int main(int argc, char* argv[]) {
                 break;  // EOF reached
             }
 
-            total_bytes_read += bytes_read;
+            total_bytes_read += static_cast<size_t>(bytes_read);
 
             // Decode from this chunk - decoder may need multiple calls per chunk
             size_t chunk_offset = 0;
@@ -104,7 +104,8 @@ int main(int argc, char* argv[]) {
                 decode_calls++;
 
                 micro_opus::OggOpusResult result =
-                    decoder.decode(input_buffer.data() + chunk_offset, bytes_read - chunk_offset,
+                    decoder.decode(input_buffer.data() + chunk_offset,
+                                   static_cast<size_t>(bytes_read) - chunk_offset,
                                    reinterpret_cast<uint8_t*>(pcm_buffer.data()),
                                    pcm_buffer.size() * sizeof(int16_t), consumed, samples);
 
@@ -192,7 +193,7 @@ int main(int argc, char* argv[]) {
                       << " audio packets)\n";
             std::cout << "Total samples written: " << wav_writer->get_samples_written() << "\n";
             std::cout << "Duration: "
-                      << (wav_writer->get_samples_written() /
+                      << (static_cast<double>(wav_writer->get_samples_written()) /
                           static_cast<double>(decoder.get_sample_rate()))
                       << " seconds\n";
             std::cout << "Output file: " << output_file << "\n";
